ajustando a componentização do conteúdo
